Print 10 violations in normal mode
Using --verbose on a large project prints out too much output
(see https://github.com/haskell/cabal/issues/10744)

Printing out all the violations if there are a lot of them is quite
slow, printing out a lot to stdout is not that fast.

Therefore we reach a compromise where we print up to 10 reasons for
failure and omit the rest. If you want all the failures you can use
--verbose.

Fixes #48